added new error state of session expired, made ui reflect it with appopriate error message

+++ MozuAndroidInStoreAssistant/app/src/main/java/com/mozu/mozuandroidinstoreassistant/app/models/authentication/UserNotAuthenticatedAuthTicket.java
@@ -10,12 +10,19 @@ public class UserNotAuthenticatedAuthTicket extends UserAuthenticationState {
 
     @Override
     public void authenticateUser() {
+        //if the auth ticket
+        if (getStateMachine().getAuthProfile() == null) {
+            getStateMachine().setCurrentUserAuthState(getStateMachine().userNotAuthenticatedNoAuthTicket);
+
+            return;
+        }
+
         //if auth ticket doesnt exist or is expired
         if (!getStateMachine().isAuthProfileStillValid()) {
             CurrentAuthTicketSerializer authTicketStore = new CurrentAuthTicketSerializer(getStateMachine().getContext());
             authTicketStore.deleteFile();
 
-            getStateMachine().setCurrentUserAuthState(getStateMachine().userNotAuthenticatedNoAuthTicket);
+            getStateMachine().setCurrentUserAuthState(getStateMachine().userAuthenticationFailedSessionExpired);
 
             return;
         }
